+++ src/client/js/scopes.js
@@ -65,7 +65,7 @@ Meteor.startup(() => {
         Tracker.autorun(() => {
             if( Roles._scopes.handle.ready()){
                 Roles._scopes.collection.find().fetchAsync().then(( fetched ) => {
-                    console.debug( 'fetched', fetched );
+                    //console.debug( 'fetched', fetched );
                     fetched.forEach(( it ) => {
                         Roles._scopes.labels.set( it._id, it.label || null );
                     })
@@ -77,5 +77,5 @@ Meteor.startup(() => {
 
 // track the scopes list
 Tracker.autorun(() => {
-    console.debug( Roles._scopes.labels.all());
+    //console.debug( Roles._scopes.labels.all());
 });